requirements.txt
frida
twisted
colorama
requests
asn1crypto
pycryptodome
